Fix bug where user couldn't delete numerical fields to replace them.
For example, if the reps field contains a 3 and user wants to change to
a 4, they would select the field, hit backspace to delete the 3, and
then type 4.  This did not work, because the reps field is a number, and
updating it with the empty string failed.

This now works correctly.  Also, there is now a check to not allow
leaving the building page until all values are filled in.